Move test to iree/samples
